+++ routerd/src/core/nexus.rs
@@ -35,6 +35,7 @@ use common::uds_server::*;
 
 use super::signal;
 use super::timer;
+use super::utils::*;
 use super::protocols::ProtocolType;
 use super::message::nexus::ProtoToNexus;
 use super::message::nexus::NexusToProto;
@@ -370,42 +371,6 @@ impl NexusConfig {
     }
 }
 
-pub fn request_parse(request: String) -> Result<(Method, String, Option<String>), CoreError> {
-    let mut lines = request.lines();
-
-    if let Some(req) = lines.next() {
-        let mut words = req.split_ascii_whitespace();
-
-        if let Some(method_str) = words.next() {
-            if let Ok(method) = Method::from_str(method_str) {
-
-                if let Some(path) = words.next() {
-                    let mut body: Option<String> = None;
-
-                    // Skip a blank line and get body if it is present.
-                    if let Some(_) = lines.next() {
-                        if let Some(b) =  lines.next() {
-                            body = Some(b.to_string());
-                        }
-                    }
-
-                    debug!("received command method: {}, path: {}, body: {:?}", method, path, body);
-
-                    Ok((method, path.to_string(), body))
-                } else {
-                    Err(CoreError::RequestInvalid(req.to_string()))
-                }
-            } else {
-                Err(CoreError::RequestInvalid(req.to_string()))
-            }
-        } else {
-            Err(CoreError::RequestInvalid(req.to_string()))
-        }
-    } else {
-        Err(CoreError::RequestInvalid("(no request line)".to_string()))
-    }
-}
-
 /// UdsServerHandler implementation for NexusConfig.
 impl UdsServerHandler for NexusConfig {
 
@@ -414,6 +379,8 @@ impl UdsServerHandler for NexusConfig {
         if let Some(request) = entry.stream_read() {
             match request_parse(request) {
                 Ok((method, path, body)) => {
+                    debug!("Received request method: {}, path: {}, body: {:?}", method, path, body);
+
                     if let Err(err) = self.handle_request(entry.index(), method, &path, body) {
                         Err(err)
                     } else {
